cleanup client usage
